Refactor MPDReducer away
These actions should not go through ReSwift.

* Remove Alamofire reference
* Upgrade ReSwift to 5
