+++ Cartfile
@@ -2,5 +2,5 @@ github "SwiftyJSON/SwiftyJSON" ~> 4.0
 github "PromiseKit/Foundation" ~> 3.0
 github "nhurden/MediaKeyTap"
 github "krzyzanowskim/CryptoSwift"
-github "ReSwift/ReSwift" "mjarvis/swift-5.0"
+github "ReSwift/ReSwift"
 github "tonyarnold/Differ"
