+++ Cartfile.resolved
@@ -1,7 +1,7 @@
-github "PromiseKit/Foundation" "3.3.2"
-github "ReSwift/ReSwift" "bc5943ad9493fc7fbad5200f690be538db8e86fb"
+github "PromiseKit/Foundation" "3.3.3"
+github "ReSwift/ReSwift" "5.0.0"
 github "SwiftyJSON/SwiftyJSON" "4.3.0"
 github "krzyzanowskim/CryptoSwift" "1.0.0"
-github "mxcl/PromiseKit" "6.8.4"
+github "mxcl/PromiseKit" "6.10.0"
 github "nhurden/MediaKeyTap" "2.2.1"
-github "tonyarnold/Differ" "1.4.2"
+github "tonyarnold/Differ" "1.4.3"
