--- Persephone/State/MPDState.swift
@@ -1,11 +1,0 @@
-//
-//  MPDState.swift
-//  Persephone
-//
-//  Created by Daniel Barber on 2019/4/30.
-//  Copyright © 2019 Dan Barber. All rights reserved.
-//
-
-import ReSwift
-
-struct MPDState: StateType {}
